requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-appconf==1.0.5
django-filter==21.1
django-imagekit==4.1.0
djangorestframework==3.12.4
drf-extra-fields==3.1.1
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.4.0
isort==5.10.1
Markdown==3.3.6
mccabe==0.6.1
pilkit==2.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
zipp==3.7.0
